binarization/requirements.txt
numpy==1.18.5
tqdm==4.47.0
opencv_python==4.4.0.44
numba==0.51.2
scipy==1.4.1
Pillow==8.0.1
pathlib
